requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
